Fix cat_sock.c : Refactored the program. The program was not correct.
I fixed some nasty copy-and-paste pieces of code.
The program created 3 orphan process without synchronisation with them
and did a mess. I fixed it to get a program tha t create a son process
which will read the file given as a positionnal parameter and
will send it to its father process through a UNIX-family socket.
The father will just retrieve data from the socket and display them.

Author : Luxon JEAN-PIERRE <luxon.jean.pierre@gmail.com>