example/ : created an example program for minimal and shared channels.
I created a small example of program that uses channels.
This program is just an modified implementation of the problem
of producer and consumers using a distributor.
This implementation will use minimal/shared channels
that should be synchronous or not.
This program can be launched in multi-thread or multi-process mode.
The multi-process mode is only available in this program, and synchronous
channel are not supported yet.