channel.c : Optimized the channel.
I wanted to replace the call to *pthread_cond_broadcast()*
with *pthread_cond_signal()*.
Unfortunately, a deadlock occurred when I launched the forward program.
So, I modified the channel structure in order to fix this bug, keeping
the call to *pthread_cond_signal()*. This modification resulted in an increase
of the efficiency of the forward program in a multi-thread and multi-process
context.

Author : Luxon JEAN-PIERRE <luxon.jean.pierre@gmail.com>