Fix channel.c : fix the condition to get a broken pipe in *channel_vsend()*.
I fixed a bug in the batch communication through channels.
The writer never get an error returned by the function when the channel is empty.
In the bad version, the function returned -1 and was ***errno*** was set
to EPIPE only if the channel is closed and empty.
So the writer could write when the channel is closed. That made the program
be incorrect.