fix channel.c : Fixes in *channel_select()* + new event - CHANNEL_EVENT_CLOSE.
**Fix**. It was not possible to get various events on a channel.
So I fixed that in order to get a set of event that should be occurred
"at the same time" in a same channel.
Consequently, it is possible to get an event related to a channel available
in reading and that should have been closed by another thread/process,
or a channel that should be written or read.
It cannot be possible to get an event related to a possible writing
if the channel is closed.

CHANNEL_EVENT_CLOSE is an event that describes the closure of a channel.
Thanks to that, it is possible to know when a channel is closed.
This event is very useful to prevent the current thread/process from writing
on a closed channel.

Some macros are also added to deal with returned events :

- CHAN_CLOSE_EVT : check if the channel was closed.
- CHAN_WRITE_EVT : check if the channel can be written.
- CHAN_READ_EVT  : check if the channel can be read.

Author : Luxon JEAN-PIERRE <luxon.jean.pierre@gmail.com>