channel.c : Refactored the channel creation.
In asynchronous channels, two condition variables must be used.
Indeed, when a group of writers and a group of readers are using
the same channel, all of them will sleep on the same condition variable,
if the operation needs to be blocked.
So it was possible to get a reader that awakes another reader
and a writer that awakes another writer. This may occurred if the number
of writers/readers is less than the capacity of the channel.

Author : Luxon JEAN-PIERRE <luxon.jean.pierre@gmail.com>